# Faster EnumCaseObjectType->describe()