--- a/src/Type/Enum/EnumCaseObjectType.php
+++ b/src/Type/Enum/EnumCaseObjectType.php
@@ -20,15 +20,11 @@ use PHPStan\Type\ObjectType;
 use PHPStan\Type\SubtractableType;
 use PHPStan\Type\Type;
 use PHPStan\Type\VerbosityLevel;
-use function array_key_exists;
 use function sprintf;
 
 /** @api */
 class EnumCaseObjectType extends ObjectType
 {
-
-	/** @var array<int, string> */
-	private array $descriptions = [];
 
 	/** @api */
 	public function __construct(
@@ -47,11 +43,9 @@ class EnumCaseObjectType extends ObjectType
 
 	public function describe(VerbosityLevel $level): string
 	{
-		$value = $level->getLevelValue();
-		if (!array_key_exists($value, $this->descriptions)) {
-			$this->descriptions[$value] = sprintf('%s::%s', parent::describe($level), $this->enumCaseName);
-		}
-		return $this->descriptions[$value];
+		$parent = parent::describe($level);
+
+		return sprintf('%s::%s', $parent, $this->enumCaseName);
 	}
 
 	public function equals(Type $type): bool