repro of https://github.com/phpstan/phpstan/issues/11263 before this PR
```
php bin/phpstan analyze slow.php --debug
  7.16s user 0.14s system 99% cpu 7.323 total

php bin/phpstan analyze slow.php --debug
  7.18s user 0.13s system 99% cpu 7.339 total
```

after this PR:
```
php bin/phpstan analyze slow.php --debug
  5.29s user 0.14s system 99% cpu 5.478 total

php bin/phpstan analyze slow.php --debug
  5.25s user 0.14s system 99% cpu 5.403 total
```