@ondrejmirtes 2024-07-03
FYI I have a local branch where the reproduction takes zero zero nothing seconds, I "just" need to make it fully work. So any further optimizations from you that shave off percents from the time are in the negligible returns territory. Stay tuned! :)