@ondrejmirtes 2024-07-03
@staabm I'm going to test it after my change. The previous PRs made sense on their own, but here we are adding caches (which means more memory consumed) so let's wait if it's still worth it.